+++ database/migrations/2024_06_21_044854_create_bookers_table.php
@@ -15,8 +15,8 @@ class CreateBookersTable extends Migration
     {
         Schema::create('bookers', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('lang_parent_id')->nullable()->default(null);
-            $table->unsignedInteger('lang_id')->nullable()->default(null);
+            $table->unsignedBigInteger('lang_parent_id')->nullable()->default(null);
+            $table->unsignedBigInteger('lang_id')->nullable()->default(null);
             $table->string('name');
             $table->text('image');
             $table->text('sale_text')->nullable();
